-"Im not sure why Kommandanter keeps trying to add -5% consumer goods to the USA..."
I have not been attempting to add such a buff, since I do not believe one is needed -  I have been simply attempting to create a niche for the US's existing "Financial Expert" political advisor, which has hung around in a state of limbo - available only after the completion of the 'Federal Housing Act' National Focus - rendering said advisor entirely useless.

I have recently made several attempts to make this advisor a tenable choice to the player - (recently I tried to make it available upon the completion of the economic focus prior to the FHA (can't remember its name) but it too was reverted.)

So - the 'Financial Expert' advisor is removed since no place exists for it.