1) Hungary, Bulgaria, and Romania can LL to Germany/Italy without any *technical* restriction imposed directly by the game blocking it.
2) The special Hungarian/Bulgarian/Romanian inf equip/art/at/aa designers are removed so that they won't be LLing broken buffed equipment to Germany...

We will have to see if it is possible to allow minor LL to Germany without them ending up being forced back into being LL minors/LL slaves. Is it possible in the jungle of competitive HOI4 for people to exercise some self control?

Some illustrative guidelines; violations of these are to be considered exploits and are against the rules:

a) You may not deliberately have LL minors who are focused on LLing majors ahead of using equipment for themselves.
b) LL from minors may not be a central part of your 'strategy,' or even a significant part of your 'strategy,' or is in any significant way planned in advance.
c) Minors may not LL equipment of a more advanced type than what another major has on their own for min maxing purposes (i.e. if Hungary LLs 1942 inf equip/artillery to Germany when Germany only has 1939 inf equip/artillery researched itself).
d) Minors may not produce an equipment type and LLing it to Germany without using most of the equipment themselves first if generally possible/reasonable (for example if Hungary produces Panthers, but doesn't have its own Panther divisions and instead just sends all its Panthers to Germany, that is not ok).
e) Minors should generally only LL to majors Germany/Italy in reasonable amounts, not giving excessive or unusual/implausible preference to LLing to majors over using the equipment themselves.

Hopefully these guidelines are sufficient that people will not abuse this, but i.e. if Germany is lacking 5k support equipment and Hungary has 2k support equip in stockpile, then Hungary can send Germany a LL of 1k support equip.

In the end though, this is necessarily vague; it is up to players to not exploit it and control yourselves. We will see if this is possible now.